landmark_frame: publish landmark based on laser line intersection
This node reads a list of known landmarks from a configuration file and
if detected publishes an appropriate landmark reference frame via tf.
Landmarks are defined in relation to intersections of lines published by
the laser_lines node. Landmark become valid in certain regions of the
map. Then, intersections detected will be compared to the reference
intersection from the configuration file in the map frame. If found and
accepted, a configurable offset is added and the proper orientation is
calculated based on a reference orientation in the map frame, and the
correspondence of one of the lines to map frame axis and its rotational
shift.

Assuming sufficiently low sensor noise (typical for laser scanners) this
gives a more stable reference frame than when relying on the
localization for proper localization in the map frame.